requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings
openai
supabase
psycopg[binary]
psycopg2-binary
SQLAlchemy
asyncpg
elasticsearch
neo4j
PyPDF2
pdfminer.six
numpy
scipy
scikit-learn
langdetect
Jinja2
